Move step_export to common.
