+++ builder/vmware/common/step_export_test.go
@@ -1,8 +1,9 @@
-package iso
+package common
 
 import (
-	"github.com/mitchellh/multistep"
 	"testing"
+
+	"github.com/mitchellh/multistep"
 )
 
 func TestStepExport_impl(t *testing.T) {
@@ -13,9 +14,9 @@ func testStepExport_wrongtype_impl(t *testing.T, remoteType string) {
 	state := testState(t)
 	step := new(StepExport)
 
-	var config Config
+	var config DriverConfig
 	config.RemoteType = "foo"
-	state.Put("config", &config)
+	state.Put("driverConfig", &config)
 
 	if action := step.Run(state); action != multistep.ActionContinue {
 		t.Fatalf("bad action: %#v", action)
